refactor(docs): MergeConflict service and controller 코드 위치 수정
